New Nodes and new Neural Network
I have train the CNN with new nodes. In the Dynamic Test folder there is a video where you can see the Neural Network working.

I will update in the future the Nodes in every folder with the name "NodeX" (in spanish is "Nodo" and i forgot to translate it jajaja)