Big update
Less time in training with this new architecture; from 3 hours/epoch to 5 min/epoch. This is possible without dropouts and smaller cnn regions.

After 90 epoch it has 99% validation and 92% test, best mark for now.

Documentation (in spanish, sorry but it s important to be in spanish) updated. It s finished, but i need to write a couple more of sentences. In the final days of May i will update the final version.

Small modifications in readme, jupyter notebooks and more things without importance.